requirements.txt
fastapi
uvicorn
torch
transformers
requests
firebase-admin
python-dotenv
google-genai